+++ packages/shared/utils/index.js
@@ -1,12 +1,7 @@
 export { uuid } from './uuid'
 export { linearEase } from './linearEase'
 export { getPathProps } from './getPathProps'
-export {
-  getWrapperStyle,
-  getTimeStyle,
-  svgStyle,
-  visuallyHidden,
-} from './styles'
+export { getWrapperStyle, getTimeStyle, visuallyHidden } from './styles'
 export { getColorsRGB, getNormalizedColors, getStroke } from './colors'
 export { colorsValidator } from './colorsValidator'
 export { getGradientId } from './getGradientId'