+++ babel.config.js
@@ -1,10 +1,7 @@
 module.exports = (api) => {
   api.cache(true)
 
-  const presets = [
-    ['@babel/preset-env', { modules: false }],
-    '@babel/preset-react',
-  ]
+  const presets = ['@babel/preset-react']
 
   const env = {
     test: {