+++ packages/shared/index.js
@@ -3,7 +3,6 @@ export {
   linearEase,
   getWrapperStyle,
   getTimeStyle,
-  svgStyle,
   getStroke,
   colorsValidator,
   visuallyHidden,